@@ -47,6 +47,7 @@
     "bulma"
   ],
   "dependencies": {
+    "@babel/polyfill": "^7.8.7",
     "@mmontes11/iot-client": "^2.0.8",
     "classnames": "^2.2.6",
     "common": "^2.0.4",
@@ -56,7 +57,7 @@
     "flag-icon-css": "^3.3.0",
     "prop-types": "^15.6.1",
     "query-string": "^6.2.0",
-    "react": "^16.3.2",
+    "react": "^16.13.0",
     "react-datepicker": "2.3.0",
     "react-dom": "^16.3.2",
     "react-google-maps": "^9.4.5",
@@ -72,14 +73,13 @@
     "socket.io-client": "^2.2.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.7.2",
+    "@babel/preset-env": "^7.7.1",
+    "@babel/preset-react": "^7.7.0",
     "autoprefixer": "^9.1.3",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^23.0.1",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-jest": "^23.4.2",
+    "babel-loader": "^8.0.6",
     "bulma": "^0.7.5",
     "coveralls": "^3.0.0",
     "css-loader": "^0.28.11",
