@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 import { compose } from "recompose";
@@ -14,55 +14,31 @@ import { injectIntl, intlShape } from "react-intl";
 import { EVENT_TYPE, MEASUREMENT_TYPE } from "constants/observationTypes";
 import { defaultTimePeriodFilter, defaultGroupBy } from "config/params";
 
-class Things extends React.Component {
-  componentDidMount() {
-    const { things, getThings } = this.props;
-    if (things.length === 0) {
-      getThings();
-    }
-  }
-
-  componentDidUpdate() {
-    const {
-      selectedThing,
-      things,
-      match: {
-        params: { thing: thingName },
-      },
-      showNotFoundError,
-    } = this.props;
-    if (selectedThing === null && things.length > 0 && thingName) {
-      const thing = things.find(t => t.name === thingName);
-      if (thing) {
-        this._selectThing(thing);
-      } else {
-        this._resetPathToRoot();
-        showNotFoundError(true);
-      }
-    }
-  }
-
-  _isSelected = thing => {
-    const { selectedThing } = this.props;
-    return selectedThing !== null && selectedThing.name === thing.name;
-  };
-
-  _selectThing = thing => {
-    const { selectThing, history } = this.props;
+const Things = ({
+  intl: { formatMessage },
+  match: {
+    params: { thing: thingName },
+  },
+  history,
+  isLoading,
+  things,
+  selectedThing,
+  shouldShowNotFoundError,
+  getThings,
+  selectThing,
+  showNotFoundError,
+}) => {
+  const _isSelected = thing => selectedThing !== null && selectedThing.name === thing.name;
+  const _resetPathToRoot = () => history.push("/things");
+  const _selectThing = thing => {
     selectThing(thing);
-    if (this._isSelected(thing)) {
-      this._resetPathToRoot();
+    if (_isSelected(thing)) {
+      _resetPathToRoot();
     } else {
       history.push(`/things/${thing.name}`);
     }
   };
-
-  _resetPathToRoot = () => {
-    const { history } = this.props;
-    history.push("/things");
-  };
-
-  _onStatsClick = (type, thing) => {
+  const _onStatsClick = (type, thing) => {
     const supportedEvents = thing.supportedObservationTypes.event;
     const supportedMeasurements = thing.supportedObservationTypes.measurement;
     let url;
@@ -72,66 +48,65 @@ class Things extends React.Component {
       url = `/stats/${type}/${supportedMeasurements[0]}?thing=${thing.name}&timePeriod=${defaultTimePeriodFilter}`;
     }
     if (url) {
-      const { history } = this.props;
       history.push(url);
     }
   };
-
-  _onDataClick = (type, observation, thing) => {
-    const { history } = this.props;
+  const _onDataClick = (type, observation, thing) =>
     history.push(
       `/data/${type}/${observation}/${defaultGroupBy}?thing=${thing.name}&timePeriod=${defaultTimePeriodFilter}`,
     );
-  };
-
-  render() {
-    const {
-      intl: { formatMessage },
-      isLoading,
-      things,
-      selectedThing,
-      shouldShowNotFoundError,
-      showNotFoundError,
-    } = this.props;
-    if (isLoading && things.length === 0) {
-      return <Loader />;
+  useEffect(() => {
+    if (things.length === 0) {
+      getThings();
     }
-    return (
-      <div>
-        <div className="container is-fluid section">
-          <div className="columns">
-            <div className="column is-one-quarter">
-              {things.map(thing => (
-                <ThingItem
-                  key={thing.name}
-                  name={thing.name}
-                  isSelected={this._isSelected(thing)}
-                  onClick={() => this._selectThing(thing)}
-                />
-              ))}
-            </div>
-            {selectedThing && (
-              <div className="column is-three-quarters">
-                <ThingDetail
-                  thing={selectedThing}
-                  onStatsClick={type => this._onStatsClick(type, selectedThing)}
-                  onDataClick={(type, observation) => this._onDataClick(type, observation, selectedThing)}
-                />
-              </div>
-            )}
-          </div>
-        </div>
-        <Modal
-          isActive={shouldShowNotFoundError}
-          onCloseClick={() => showNotFoundError(false)}
-          messageStyle="is-danger"
-          title={formatMessage({ id: "Error" })}
-          subTitle={formatMessage({ id: "Thing not found" })}
-        />
-      </div>
-    );
+    if (selectedThing === null && things.length > 0 && thingName) {
+      const thing = things.find(t => t.name === thingName);
+      if (thing) {
+        _selectThing(thing);
+      } else {
+        _resetPathToRoot();
+        showNotFoundError(true);
+      }
+    }
+  }, [things]);
+  if (isLoading && things.length === 0) {
+    return <Loader />;
   }
-}
+  return (
+    <div>
+      <div className="container is-fluid section">
+        <div className="columns">
+          <div className="column is-one-quarter">
+            {things.map(thing => (
+              <ThingItem
+                key={thing.name}
+                name={thing.name}
+                isSelected={_isSelected(thing)}
+                onClick={() => _selectThing(thing)}
+              />
+            ))}
+          </div>
+          {selectedThing && (
+            <div className="column is-three-quarters">
+              <ThingDetail
+                thing={selectedThing}
+                onStatsClick={type => _onStatsClick(type, selectedThing)}
+                onDataClick={(type, observation) => _onDataClick(type, observation, selectedThing)}
+              />
+            </div>
+          )}
+        </div>
+      </div>
+      <Modal
+        isActive={shouldShowNotFoundError}
+        onCloseClick={() => showNotFoundError(false)}
+        messageStyle="is-danger"
+        title={formatMessage({ id: "Error" })}
+        subTitle={formatMessage({ id: "Thing not found" })}
+      />
+    </div>
+  );
+};
 
 Things.propTypes = {
   intl: intlShape.isRequired,
